requirements.txt
Flask==0.10.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-Cache==0.13.1
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.7
Werkzeug==0.9.6
itsdangerous==0.24
GeoAlchemy2==0.2.4
geojson==1.0.9
Shapely==1.4.3
PyYAML==3.11

# >= so that it wont complain if a later version is installer with apt-get
psycopg2>=2.5.4
numpy>=1.9.0